move language elements into templater.language
- except tokens
- update imports with moves
- templater.language classes are all public
- made many non final fields final
