+++ src/templater/compiler/parser/TokenMatchable.java
@@ -2,9 +2,7 @@ package templater.compiler.parser;
 
 import templater.compiler.Matchable;
 
-abstract class TokenMatchable<T> implements
-  Matchable<TokenQueue.Iterator, T> {
-
+abstract class TokenMatchable<T> implements Matchable<TokenQueue.Iterator, T> {
   @Override
   public final T tryMatch(TokenQueue.Iterator input) {
     input.mark();
