+++ src/templater/language/AttributeElement.java
@@ -1,0 +1,21 @@
+package templater.language;
+
+import templater.compiler.LanguageElement;
+
+public class AttributeElement extends LanguageElement {
+  private final String key;
+  private final StringResolvables value;
+
+  public AttributeElement(String key, StringResolvables value) {
+    this.key = key;
+    this.value = value;
+  }
+
+  public String getKey() {
+    return this.key;
+  }
+
+  public StringResolvables getValue() {
+    return this.value;
+  }
+}
